Schedule market data pipeline with airflow, Add ccxt integration for both market data and broker

+++ broker/percent_fee_model.py
@@ -1,4 +1,4 @@
-from broker.fee_model import FeeModel
+from trazy_analysis.broker.fee_model import FeeModel
 
 
 class PercentFeeModel(FeeModel):
@@ -48,7 +48,7 @@ class PercentFeeModel(FeeModel):
         return self.commission_pct * abs(consideration)
 
     def calc_max_size_for_cash(self, cash: float, price: float) -> int:
-        return int(cash / (price * (1 + self.commission_pct)))
+        return cash / (price * (1 + self.commission_pct))
 
     def _calc_tax(
         self, symbol: str, size: int, consideration: float, broker: "Broker" = None
